Move populateConvContext to a source file.
